+++ src/frontend/sider/Fagsak/Behandling/sider/Vedtak/Vedtaksperioder/utils.ts
